CSV to JSON Overhaul, minor multi-trait req support
Redid significant parts of app/data to allow for converting the awkward CSV format to a more flexible json.

Starting redoing traits so they support multiple requests.
