+++ app/data/__main__.py
@@ -1,0 +1,3 @@
+from . import csvtojson
+
+csvtojson.main()
